requirements.txt
streamlit
google-cloud-storage
google-auth
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
Pillow
moviepy==1.0.3
decorator==4.4.2
google-cloud-aiplatform>=1.38.0 # For Vertex AI Gemini
